Rename the controller and delete oam_dma_transfer
