+++ lib_gb/src/mmu/mod.rs
@@ -5,8 +5,7 @@ pub mod vram;
 pub mod io_ports;
 pub mod carts;
 pub mod access_bus;
-pub mod oam_dma_transfer;
 pub mod io_bus;
 pub mod interrupts_handler;
 pub mod external_memory_bus;
-pub mod dma_controller;
+pub mod oam_dma_controller;
